Addition, negatives, complement values
# Functional changes
Reliable custom addition of numeric values in different bases / with a negative polarity.
Digit inherits from BaseDigit which is used when returning any two's complement or other non-place-value numeric values.

# Refactoring
Used some @decorated methods to reduce the number of calls to constructor functions. In this way, re-used part of the addition function to add 1 when "normalising" numbers. This may be unnecessarily complex as static methods could equally be represented as a function, which might be more consistent.