Rewrite instructions and option names
Sensible instructions have been added to the index.html page. Renamed calculations to be more consistent with how they are displayed in the list of steps and renamed buttons to "Create New -".

A list of steps ending in a (hanging) addition sign will cause an alert on the client-side. Capital letters are now accepted in hexadecimal numbers (i.e. 1F).